relacionamentos definidos. realizar os relacionamentos no código nas entidades e efcore

+++ Lojinha.Shared.Models/Devedor.cs
@@ -8,8 +8,8 @@ namespace Lojinha.Shared.Models
         //adicionar id cliente
         // adicionar id venda
         
-        public double DividaInicial { get; set; }
-        public double DividaAtual { get; set; }
+        public decimal DividaInicial { get; set; }
+        public decimal DividaAtual { get; set; }
         public DateTime DataInicio { get; set; }
         public DateTime DataAtualizacao { get; set; }
         public DateTime? DataFim { get; set; }
